restructure section 2 codebase
